azblob: use `copy` to implement `read_to_end`
