+++ components/cloud/Cargo.toml
@@ -21,3 +21,5 @@ url = "2.0"
 uuid = { version = "0.8", features = ["v4"] }
 
 [dev-dependencies]
+pin-project = "1"
+tokio = { version = "1.5", features = ["rt"] }
