+++ components/cloud/gcp/Cargo.toml
@@ -30,5 +30,3 @@ url = "2.0"
 
 [dev-dependencies]
 matches = "0.1.8"
-pin-project = "1"
-tokio = { version = "1.5", features = ["rt"] }
