+++ components/cloud/gcp/src/lib.rs
@@ -1,5 +1,4 @@
 // Copyright 2021 TiKV Project Authors. Licensed under Apache-2.0.
-#![feature(test)]
 
 #[macro_use]
 extern crate slog_global;
